fix gpfdist crash
Long log:
The bug is strongly related to special data and it will be triggered only if there is no extra space on the disk.
We failed to reproduce this bug and can only analyze it by gdb debugging. After analyzing the core dump, we found fd->u.z->s.state=0x0 in gz_file_write_one_chunk(). This may be the reason.
The solution is to return -1 as long as s.state=0x0.

Signed-off-by: Yongtao Huang <yongtaoh@vmware.com>